functions/testfunc/requirements.txt
functions-framework==3.*
google-cloud-logging==3.*
